+++ packages/expo-facebook/ios/EXFacebook/EXFacebook.m
@@ -14,27 +14,21 @@ NSString * const EXFacebookLoginErrorDomain = @"ERR_FACEBOOK_LOGIN";
 
 UM_EXPORT_MODULE(ExponentFacebook)
 
+UM_EXPORT_METHOD_AS(setAdvertiserTrackingEnabledAsync,
+                    setAdvertiserTrackingEnabled:(BOOL)enabled
+                    resolver:(UMPromiseResolveBlock)resolve
+                    rejecter:(UMPromiseRejectBlock)reject)
+{
+  BOOL result = [FBSDKSettings setAdvertiserTrackingEnabled:enabled];
+  resolve(@(result));
+}
+
 UM_EXPORT_METHOD_AS(setAutoLogAppEventsEnabledAsync,
                     setAutoLogAppEventsEnabled:(BOOL)enabled
                     resolver:(UMPromiseResolveBlock)resolve
                     rejecter:(UMPromiseRejectBlock)reject)
 {
   [FBSDKSettings setAutoLogAppEventsEnabled:enabled];
-  resolve(nil);
-}
-
-UM_EXPORT_METHOD_AS(setAutoInitEnabledAsync,
-                    setAutoInitEnabled:(BOOL)enabled
-                    resolver:(UMPromiseResolveBlock)resolve
-                    rejecter:(UMPromiseRejectBlock)reject)
-{
-  // If enabled is true, the line below will initialize the SDK.
-  // This behavior is different than on Android where one needs
-  // to initialize the SDK explicitly. We have no power over this,
-  // and to mitigate this difference we will NOT add initializing
-  // to the respective method on Android, but we will instruct users
-  // to initialize the SDK manually on both platforms instead.
-  [FBSDKSettings setAutoInitEnabled:enabled];
   resolve(nil);
 }